# docs/swagger.yaml
basePath: /
definitions:
  models.Post:
    properties:
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
    type: object
  models.User:
    properties:
      dob:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: A simple RESTful API built with Go and PostgreSQL.
  title: Basic Web API
  version: "1.0"
paths:
  /404:
    get:
      description: Returns a 404 Not Found error
      produces:
      - application/json
      responses:
        "404":
          description: Not Found
          schema:
            type: string
      summary: Handle 404 errors
      tags:
      - errors
  /posts:
    get:
      consumes:
      - application/json
      description: Get all posts or create a new post
      parameters:
      - description: Post object for creation
        in: body
        name: post
        schema:
          $ref: '#/definitions/models.Post'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Post'
            type: array
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Post'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Handle posts
      tags:
      - posts
  /users:
    get:
      consumes:
      - application/json
      description: Get all users or create a new user
      parameters:
      - description: User object for creation
        in: body
        name: user
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Handle users
      tags:
      - users
swagger: "2.0"